rename logic gates test
