+++ pyproject.toml
@@ -34,5 +34,5 @@ env_list = [
 description = "Run tests under {base_python}"
 change_dir="tests"
 commands = [
-  ["python", "test_and_or.py"],
+  ["python", "test_logic_gates.py"],
 ]
